+++ src/astronomy/mod.rs
@@ -2,3 +2,4 @@ pub mod orbit;
 pub mod planet;
 pub mod planetary_system;
 pub mod star;
+pub mod starfield;