+++ src/resources/mod.rs
@@ -1,3 +1,4 @@
 pub mod assets;
+pub mod camera;
 pub mod game_time;
 pub mod state;